Converted int64_t index to int32_t
int64_t is not supported by hammerblade

+++ hammerblade/torch/kernel/kernel_index_add.cpp
@@ -32,10 +32,10 @@ extern "C" {
   }
 
 
-  int64_t get_element_index(HBTensor<float> &ten, int add_dim, int index, int64_t elementInSlice) {
-      int64_t offset = 0;
+  int get_element_index(HBTensor<float> &ten, int add_dim, int index, int elementInSlice) {
+      int offset = 0;
       for (int i = ten.ndim()-1; i > 0; --i) {
-          int32_t size = (i == add_dim)? 1 : ten.dim(i);
+          int size = (i == add_dim)? 1 : ten.dim(i);
           offset += (elementInSlice % size) * ten.stride(i);
           elementInSlice /= size;
       }
@@ -51,14 +51,14 @@ extern "C" {
           hb_tensor_t* t0_p,
           hb_tensor_t* t1_p,
           hb_tensor_t* t2_p,
-          int64_t* dim_p,
-          int64_t* sliceSize_p) {
+          int32_t* dim_p,
+          int32_t* sliceSize_p) {
 
     auto dst = HBTensor<float>(t0_p);
     auto src = HBTensor<float>(t1_p);
-    auto idx = HBTensor<int64_t>(t2_p);
-    int64_t dim = *dim_p;
-    int64_t sliceSize = *sliceSize_p;
+    auto idx = HBTensor<int32_t>(t2_p);
+    int dim = *dim_p;
+    int sliceSize = *sliceSize_p;
 
     // Start profiling
     bsg_cuda_print_stat_kernel_start();
@@ -67,13 +67,13 @@ extern "C" {
     init_locks();
     g_barrier.sync();
 
-    for (int64_t srcIndex = 0; srcIndex < idx.numel(); ++srcIndex) {
-        int64_t dstIndex = idx(srcIndex);
-        for (int64_t linearIndex = bsg_id; linearIndex < sliceSize; linearIndex += BSG_TILE_GROUP_X_DIM * BSG_TILE_GROUP_Y_DIM) {
-            int64_t dst_element_idx = get_element_index(dst, dim, dstIndex, linearIndex);
-            int64_t src_element_idx = get_element_index(src, dim, srcIndex, linearIndex);
+    for (int srcIndex = 0; srcIndex < idx.numel(); ++srcIndex) {
+        int dstIndex = idx(srcIndex);
+        for (int linearIndex = bsg_id; linearIndex < sliceSize; linearIndex += BSG_TILE_GROUP_X_DIM * BSG_TILE_GROUP_Y_DIM) {
+            int dst_element_idx = get_element_index(dst, dim, dstIndex, linearIndex);
+            int src_element_idx = get_element_index(src, dim, srcIndex, linearIndex);
 
-            int64_t dst_lock_idx = dst_element_idx && 0xFF;
+            int dst_lock_idx = dst_element_idx && 0xFF;
             int *dst_lock = &lock[dst_lock_idx];
 
             aquire_lock(dst_lock);
@@ -95,18 +95,18 @@ extern "C" {
           hb_tensor_t* t0_p,
           hb_tensor_t* t1_p,
           hb_tensor_t* t2_p,
-          int64_t* dim_p,
-          int64_t* sliceSize_p,
-          int64_t* numIndices_p,
+          int32_t* dim_p,
+          int32_t* sliceSize_p,
+          int32_t* numIndices_p,
           int32_t* indexMajorMode_p) {
 
     auto dst = HBTensor<float>(t0_p);
     auto src = HBTensor<float>(t1_p);
-    auto idx = HBTensor<int64_t>(t2_p);
-    int64_t dim = *dim_p;
-    int64_t sliceSize = *sliceSize_p;
-    int64_t numIndices = *numIndices_p;
-    int32_t indexMajorMode = *indexMajorMode_p;
+    auto idx = HBTensor<int32_t>(t2_p);
+    int dim = *dim_p;
+    int sliceSize = *sliceSize_p;
+    int numIndices = *numIndices_p;
+    int indexMajorMode = *indexMajorMode_p;
 
       // Start profiling
     bsg_cuda_print_stat_kernel_start();
@@ -117,7 +117,7 @@ extern "C" {
 
 
     for (int linearIndex = bsg_id; linearIndex < src.numel(); linearIndex += BSG_TILE_GROUP_X_DIM * BSG_TILE_GROUP_Y_DIM) {
-        int64_t srcIndex, elementInSlice;
+        int srcIndex, elementInSlice;
         if (indexMajorMode == 1) {
             srcIndex = linearIndex / sliceSize;
             elementInSlice = linearIndex % sliceSize;
@@ -126,12 +126,12 @@ extern "C" {
             elementInSlice = linearIndex / numIndices;
         }
         //bsg_printf("tile %d, srcIndex: %d, elementInSlice: %d\n", bsg_id, srcIndex, elementInSlice);
-        int64_t dstIndex = idx(srcIndex);
+        int dstIndex = idx(srcIndex);
 
-        int64_t dst_element_idx = get_element_index(dst, dim, dstIndex, elementInSlice);
-        int64_t src_element_idx = get_element_index(src, dim, srcIndex, elementInSlice);
+        int dst_element_idx = get_element_index(dst, dim, dstIndex, elementInSlice);
+        int src_element_idx = get_element_index(src, dim, srcIndex, elementInSlice);
 
-        int64_t dst_lock_idx = dst_element_idx && 0xFF;
+        int dst_lock_idx = dst_element_idx && 0xFF;
         int *dst_lock = &lock[dst_lock_idx];
 
         aquire_lock(dst_lock);
@@ -149,9 +149,9 @@ extern "C" {
   }
 
   HB_EMUL_REG_KERNEL(tensorlib_index_add_small_index, hb_tensor_t*, hb_tensor_t*, hb_tensor_t*,
-                     int64_t*, int64_t*)
+                     int32_t*, int32_t*)
   HB_EMUL_REG_KERNEL(tensorlib_index_add_large_index, hb_tensor_t*, hb_tensor_t*, hb_tensor_t*,
-                     int64_t*, int64_t*, int64_t*, int32_t*)
+                     int32_t*, int32_t*, int32_t*, int32_t*)
 
 }
